Merge

See merge request informatik/qe/sepsss20/g4t4!67

+++ server/src/main/java/uibk/ac/at/prodiga/utils/JwtTokenUtil.java
@@ -15,7 +15,7 @@ import java.util.function.Function;
 public class JwtTokenUtil implements Serializable {
 
     private static final long serialVersionUID = 3482234902843298432L;
-    private static final long JWT_TOKEN_VALIDITY = 3 * 60 * 60; // 3 hours
+    private static final long JWT_TOKEN_VALIDITY = (long)3 * 60 * 60; // 3 hours
 
     @Value("${jwt.secret}")
     private String secret;
